Move i18n.expand to separate module
